initializing input keys

+++ Switch.h
@@ -12,7 +12,7 @@
 void Switch_Init(void);
 
 // return current state of switches
-uint32_t Switch_In(void);
+uint8_t Switch_In(void);
 
 
 #endif /* SWITCH_H_ */
